Setup Travis/Coveralls: - Added both .travis.yml and .coveralls.yml - Restructure folders
