+++ .coveralls.yml
@@ -1,0 +1,2 @@
+service_name: travis-pro
+repo_token: K3qXgjtrZZWwRd9EjwWI9vdU3Jp0EgQDK
