+++ .travis.yml
@@ -1,0 +1,20 @@
+language: node_js
+
+node_js:
+  - '10'
+
+sudo: false
+
+before_install:
+  - curl -o- -L https://yarnpkg.com/install.sh | bash
+  - export PATH="$HOME/.yarn/bin:$PATH"
+
+cache:
+  yarn: true
+  directories:
+    - '.eslintcache'
+    - 'node_modules'
+
+script:
+  - yarn test --updateSnapshot && cat ./coverage/lcov.info | coveralls
+
