+++ package.json
@@ -13,7 +13,7 @@
     "lint": "eslint --quiet",
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
+    "test": "react-scripts test --env=jsdom --coverage",
     "eject": "react-scripts eject"
   },
   "author": "Seunkoko",
@@ -23,6 +23,7 @@
   },
   "devDependencies": {
     "babel-eslint": "^10.0.1",
+    "coveralls": "^3.0.2",
     "eslint": "^5.6.1",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-plugin-filenames": "^1.3.2",
